+++ pyonsnalcolor-api/src/main/java/com/pyonsnalcolor/auth/controller/MemberController.java
@@ -18,7 +18,7 @@ import org.springframework.web.bind.annotation.*;
 import javax.validation.Valid;
 
 @Tag(name = "(인증 제외)사용자 관련 api")
-@Controller
+@RestController
 @RequiredArgsConstructor
 @RequestMapping("/member")
 public class MemberController {
