+++ pyonsnalcolor-api/src/main/java/com/pyonsnalcolor/push/controller/PushRecordController.java
@@ -1,0 +1,13 @@
+package com.pyonsnalcolor.push.controller;
+
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequestMapping("/push/record")
+public class PushRecordController {
+
+    // 기록 조회
+
+    // 기록 등록
+}
