+++ pyonsnalcolor-api/src/main/java/com/pyonsnalcolor/push/service/PushRecordService.java
@@ -1,0 +1,13 @@
+package com.pyonsnalcolor.push.service;
+
+import com.pyonsnalcolor.push.repository.PushKeywordRepository;
+import com.pyonsnalcolor.push.repository.PushRecordRepository;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+public class PushRecordService {
+
+    @Autowired
+    PushRecordRepository pushRecordRepository;
+}
