+++ pyonsnalcolor-domain/src/main/java/com/pyonsnalcolor/push/PushKeyword.java
@@ -1,32 +1,25 @@
-package com.pyonsnalcolor.alarm;
+package com.pyonsnalcolor.push;
 
 import com.pyonsnalcolor.member.Member;
-import com.pyonsnalcolor.product.enumtype.StoreType;
 import lombok.*;
 
 import javax.persistence.*;
-import java.time.LocalDateTime;
+import javax.validation.constraints.Pattern;
 
 @Builder
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-@Table(name = "push_record")
-public class PushRecord {
+@Table(name = "push_keyword")
+public class PushKeyword {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "push_record_id")
+    @Column(name = "push_keyword_id")
     private Long id;
 
-    private String title;
-
-    private String body;
-
-    private StoreType storeType;
-
-    @Column(name = "created_time")
-    private LocalDateTime createdTime;
+    @Pattern(regexp="^[0-9a-zA-Zㄱ-ㅎ가-힣]{1,10}")
+    private String name;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
