+++ pyonsnalcolor-domain/src/main/java/com/pyonsnalcolor/push/PushRecord.java
@@ -1,25 +1,32 @@
-package com.pyonsnalcolor.alarm;
+package com.pyonsnalcolor.push;
 
 import com.pyonsnalcolor.member.Member;
+import com.pyonsnalcolor.product.enumtype.StoreType;
 import lombok.*;
 
 import javax.persistence.*;
-import javax.validation.constraints.Pattern;
+import java.time.LocalDateTime;
 
 @Builder
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
 @Entity
-@Table(name = "push_keyword")
-public class PushKeyword {
+@Table(name = "push_record")
+public class PushRecord {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "push_keyword_id")
+    @Column(name = "push_record_id")
     private Long id;
 
-    @Pattern(regexp="^[0-9a-zA-Zㄱ-ㅎ가-힣]{1,10}")
-    private String name;
+    private String title;
+
+    private String body;
+
+    private StoreType storeType;
+
+    @Column(name = "created_time")
+    private LocalDateTime createdTime;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "member_id")
